+++ src/analytics/process_pattern_matcher.cpp
@@ -185,7 +185,7 @@ double ProcessPatternMatcher::computeGraphSimilarity(
                           trace_set.begin(), trace_set.end(),
                           std::inserter(intersection_set, intersection_set.begin()));
     size_t sym_diff = (pat_set.size() + trace_set.size()) - 2 * intersection_set.size();
-    double denom = pat_set.size() + trace_set.size();
+    double denom = static_cast<double>(pat_set.size() + trace_set.size());
     double edit_norm = (denom > 0) ? 1.0 - static_cast<double>(sym_diff) / denom : 1.0;
 
     return 0.30 * node_jac + 0.30 * edge_jac + 0.25 * path_sim + 0.15 * edit_norm;
@@ -384,7 +384,7 @@ ProcessPatternMatcher::findSimilar(
                               trace_set.begin(), trace_set.end(),
                               std::inserter(inter_set, inter_set.begin()));
         size_t sym_diff = (pat_set.size() + trace_set.size()) - 2 * inter_set.size();
-        double den = pat_set.size() + trace_set.size();
+        double den = static_cast<double>(pat_set.size() + trace_set.size());
         metrics.edit_distance = (den > 0) ? static_cast<double>(sym_diff) / den : 0.0;
 
         // Matched / missing / extra activities
