+++ src/security/manifest_signer.cpp
@@ -42,7 +42,7 @@ namespace {
         bio = BIO_push(b64, bio);
         
         BIO_set_flags(bio, BIO_FLAGS_BASE64_NO_NL);
-        BIO_write(bio, data.data(), data.size());
+        BIO_write(bio, data.data(), static_cast<int>(data.size()));
         BIO_flush(bio);
         BIO_get_mem_ptr(bio, &buffer_ptr);
         
@@ -56,10 +56,10 @@ namespace {
     std::vector<uint8_t> base64Decode(const std::string& encoded) {
         BIO *bio, *b64;
         
-        int decode_len = encoded.length();
+        int decode_len = static_cast<int>(encoded.length());
         std::vector<uint8_t> result(decode_len);
         
-        bio = BIO_new_mem_buf(encoded.data(), encoded.length());
+        bio = BIO_new_mem_buf(encoded.data(), static_cast<int>(encoded.length()));
         b64 = BIO_new(BIO_f_base64());
         bio = BIO_push(b64, bio);
